rename: 회비 납부 관련 컴포넌트 PaymentStatus 폴더로 이동
